% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synopsis.R
\name{synopsis}
\alias{synopsis}
\title{Check and import a dataset, and provide a summary after importing the dataset}
\arguments{
\item{file_path}{dataset or path to file}
}
\value{
import and summary of dataset
}
\description{
Check and import a dataset, and provide a summary after importing the dataset
}
\author{
Naresh de Bruijn
}
