% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synopsis.R
\name{synopsis}
\alias{synopsis}
\title{Check and import a dataset, and provide a summary after importing the dataset}
\arguments{
\item{file_path}{dataset or path to file}
}
\value{
summary of the imported data
}
\description{
Check and import a dataset, and provide a summary after importing the dataset
}
\examples{

# results <- synopsis("./data/species.csv")
# print(results$data)
# print(results$class)
# print(results$dim)
# print(results$str)
# print(results$typeof)
# print(results$names)
# print(results$NA_info)
# print(resluts)

}
\author{
Naresh de Bruijn
}
